Se liberan las instancias de los autos que estan en funcionamiento, los que estan estacionados no
