+++ Source/PositionManager.cs
@@ -117,21 +117,6 @@ namespace ODStudy
             return (goalPosition - busPosition).magnitude <= 40f;
         }
 
-        public void cleanTraffic()
-        {
-            VehicleManager instance = Singleton<VehicleManager>.instance;
-            uint totalVehicles = instance.m_vehicles.m_size; ;
-
-            if (totalVehicles == 0)
-            {
-                return;
-            }
-
-            for (ushort i = 1; i < totalVehicles; i++)
-            {
-                instance.ReleaseVehicle(i);
-            }
-        }
 
         public void reRespawn(ushort vehicleID)
         {
